CPM-1937 demo command add isProductionEnv check

+++ app/Console/Commands/UPG0506Demo.php
@@ -75,51 +75,53 @@ class UPG0506Demo extends Command
 
     private function clearTestData()
     {
-        $ccdas = Ccda::where('mrn', '334417')
-            ->get();
+        if (! isProductionEnv()){
+            $ccdas = Ccda::where('mrn', '334417')
+                         ->get();
 
-        foreach ($ccdas as $ccda) {
-            $ccda->media()
-                ->get()
-                ->each(function ($media) {
-                $media->delete();
-            });
+            foreach ($ccdas as $ccda) {
+                $ccda->media()
+                     ->get()
+                     ->each(function ($media) {
+                         $media->delete();
+                     });
 
-            \DB::table('media')
-                      ->where('custom_properties->is_pdf', 'true')
-                      ->where('custom_properties->is_upg0506', 'true')
-                      ->where('custom_properties->care_plan->demographics->mrn_number', '334417')
-                      ->delete();
+                \DB::table('media')
+                   ->where('custom_properties->is_pdf', 'true')
+                   ->where('custom_properties->is_upg0506', 'true')
+                   ->where('custom_properties->care_plan->demographics->mrn_number', '334417')
+                   ->delete();
 
-            $dm = $ccda->directMessage()->first();
+                $dm = $ccda->directMessage()->first();
 
-            if ($dm) {
-                $dm->media()
-                    ->get()
-                    ->each(function ($media) {
-                                $media->delete();
-                            });
-                $dm->delete();
+                if ($dm) {
+                    $dm->media()
+                       ->get()
+                       ->each(function ($media) {
+                           $media->delete();
+                       });
+                    $dm->delete();
+                }
+
+                $ccda->importedMedicalRecord()->forceDelete();
+
+                $patient = $ccda->getPatient();
+
+                if ($patient){
+                    $patient->patientSummaries()->delete();
+                    $patient->forceDelete();
+                }
+
+                $ccda->forceDelete();
             }
 
-            $ccda->importedMedicalRecord()->forceDelete();
+            $pdf = Media::where('custom_properties->is_upg0506', 'true')
+                        ->where('custom_properties->care_plan->demographics->mrn_number', '334417')
+                        ->first();
 
-            $patient = $ccda->getPatient();
-
-            if ($patient){
-                $patient->patientSummaries()->delete();
-                $patient->forceDelete();
+            if ($pdf){
+                $pdf->delete();
             }
-
-            $ccda->forceDelete();
-        }
-
-        $pdf = Media::where('custom_properties->is_upg0506', 'true')
-             ->where('custom_properties->care_plan->demographics->mrn_number', '334417')
-            ->first();
-
-        if ($pdf){
-            $pdf->delete();
         }
 
         User::whereFirstName('Barbara')
